rename start cmd folder
