+++ scopes/ui-foundation/ui/start-cmd/cli-output.tsx
