+++ scopes/ui-foundation/ui/start-cmd/output-templates/clear-console.tsx
