+++ scopes/ui-foundation/ui/start-cmd/output-templates/component-change.tsx
