+++ scopes/ui-foundation/ui/start-cmd/output-templates/component-preview-server-started-headers.tsx
