+++ scopes/ui-foundation/ui/start-cmd/output-templates/index.ts
