+++ scopes/ui-foundation/ui/start-cmd/output-templates/ink-spinner.tsx
