+++ scopes/ui-foundation/ui/start-cmd/output-templates/standalone-new-line.tsx
