+++ scopes/ui-foundation/ui/start-cmd/output-templates/starting-main-ui-server.tsx
