+++ scopes/ui-foundation/ui/start-cmd/output-templates/starting.tsx
