+++ scopes/ui-foundation/ui/start-cmd/output-templates/ts-errors.tsx
