+++ scopes/ui-foundation/ui/start-cmd/output-templates/ui-servers-are-ready.tsx
