+++ scopes/ui-foundation/ui/start-cmd/output-templates/ui-servers-is-ready-in-scope.tsx
